+++ client/src/screens/Game.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import Combat from '../components/Dice/Combat';
 import RollDice from '../components/Dice/RollDice';
 
 export default function Game() {
@@ -7,7 +8,7 @@ export default function Game() {
         <img className='w-screen' src="https://upload.wikimedia.org/wikipedia/commons/thumb/5/5d/Risk_game_board.svg/600px-Risk_game_board.svg.png" alt="risk" />
         <div className='flex p-5'>
           <RollDice/>
-          <RollDice/>
+          <Combat troopNum1={3} troopNum2={2}/>
         </div>
     </div>
   );
